# My_Warehouse/api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateComingProducts:
    properties:
      amount:
        type: integer
      barcode:
        type: string
      categroy_id:
        type: string
      coming_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  models.UpdateComingProducts:
    properties:
      amount:
        type: integer
      barcode:
        type: string
      categroy_id:
        type: string
      coming_id:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /comingproducts:
    get:
      consumes:
      - application/json
      description: Get List ComingProducts
      operationId: get_list_comingProducts
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search_by_category
        in: query
        name: SearchByCategory
        type: string
      - description: search_by_barcode
        in: query
        name: SearchByBarcode
        type: string
      - description: search_by_coming_id
        in: query
        name: SearchByComingID
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List ComingProducts
      tags:
      - ComingProducts
    post:
      consumes:
      - application/json
      description: Create ComingProducts
      operationId: create_comingProducts
      parameters:
      - description: CreateComingProductsRequest
        in: body
        name: comingProducts
        required: true
        schema:
          $ref: '#/definitions/models.CreateComingProducts'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ComingProducts
      tags:
      - ComingProducts
  /comingproducts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete ComingProducts
      operationId: delete_comingProducts
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete ComingProducts
      tags:
      - ComingProducts
    get:
      consumes:
      - application/json
      description: Get By ID ComingProducts
      operationId: get_by_id_comingProducts
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID ComingProducts
      tags:
      - ComingProducts
    put:
      consumes:
      - application/json
      description: Update ComingProducts
      operationId: update_comingProducts
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateComingProductsRequest
        in: body
        name: comingProducts
        required: true
        schema:
          $ref: '#/definitions/models.UpdateComingProducts'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ComingProducts
      tags:
      - ComingProducts
  /do-income/{coming_id}:
    post:
      consumes:
      - application/json
      description: Send Coming
      operationId: Income_Warehouse_
      parameters:
      - description: coming_id
        in: path
        name: coming_id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Send Coming
      tags:
      - Send Coming
...
swagger: "2.0"